checkpoint for network tape
SCCS-vsn: usr.sbin/amd/amd/amd.c 5.3
SCCS-vsn: usr.sbin/amd/amd/nfs_start.c 5.3
SCCS-vsn: usr.sbin/amd/amd/nfs_subr.c 5.3
SCCS-vsn: usr.sbin/amd/amd/clock.c 5.3
SCCS-vsn: usr.sbin/amd/amd/efs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/amq_subr.c 5.3
SCCS-vsn: usr.sbin/amd/amd/ifs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/host_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/map.c 5.3
SCCS-vsn: usr.sbin/amd/amd/get_args.c 5.3
SCCS-vsn: usr.sbin/amd/amd/info_hes.c 5.3
SCCS-vsn: usr.sbin/amd/amd/info_ndbm.c 5.3
SCCS-vsn: usr.sbin/amd/amd/info_nis.c 5.3
SCCS-vsn: usr.sbin/amd/amd/info_passwd.c 5.3
SCCS-vsn: usr.sbin/amd/amd/mapc.c 5.3
SCCS-vsn: usr.sbin/amd/amd/info_file.c 5.3
SCCS-vsn: usr.sbin/amd/amd/mntfs.c 5.3
SCCS-vsn: usr.sbin/amd/amd/mtab.c 5.3
SCCS-vsn: usr.sbin/amd/amd/misc_rpc.c 5.3
SCCS-vsn: usr.sbin/amd/amd/nfs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/mount_fs.c 5.3
SCCS-vsn: usr.sbin/amd/amd/opts.c 5.3
SCCS-vsn: usr.sbin/amd/amd/pfs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/restart.c 5.3
SCCS-vsn: usr.sbin/amd/amd/amd.8 5.4
SCCS-vsn: usr.sbin/amd/amd/rpc_fwd.c 5.3
SCCS-vsn: usr.sbin/amd/amd/sched.c 5.3
SCCS-vsn: usr.sbin/amd/amd/sfs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/ufs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/srvr_afs.c 5.3
SCCS-vsn: usr.sbin/amd/amd/srvr_nfs.c 5.3
SCCS-vsn: usr.sbin/amd/amd/util.c 5.3
SCCS-vsn: usr.sbin/amd/amd/umount_fs.c 5.3
SCCS-vsn: usr.sbin/amd/amd/info_union.c 5.3
SCCS-vsn: usr.sbin/amd/amd/nfsx_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/am_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/afs_ops.c 5.3
SCCS-vsn: usr.sbin/amd/amd/xutil.c 5.3
SCCS-vsn: usr.sbin/amd/amd/wire.c 5.3
SCCS-vsn: usr.sbin/amd/amd/Makefile 5.4
SCCS-vsn: usr.sbin/amd/amd/ChangeLog 1.3
SCCS-vsn: usr.sbin/amd/doc/amdref.cps 1.2
SCCS-vsn: usr.sbin/amd/doc/amdref.fns 1.2
SCCS-vsn: usr.sbin/amd/doc/amdref.kys 1.2
SCCS-vsn: usr.sbin/amd/doc/amdref.pgs 1.2
SCCS-vsn: usr.sbin/amd/doc/amdref.texinfo 1.3
SCCS-vsn: usr.sbin/amd/doc/amdref.tps 1.2
SCCS-vsn: usr.sbin/amd/doc/amdref.vrs 1.2
SCCS-vsn: usr.sbin/amd/doc/texinfo.tex 1.2
SCCS-vsn: usr.sbin/amd/doc/Makefile 5.3
SCCS-vsn: usr.sbin/amd/rpcx/amq_svc.c 5.3
SCCS-vsn: usr.sbin/amd/rpcx/amq_xdr.c 5.3
SCCS-vsn: usr.sbin/amd/rpcx/amq_clnt.c 5.3
SCCS-vsn: usr.sbin/amd/rpcx/mount.h 5.3
SCCS-vsn: usr.sbin/amd/rpcx/mount_xdr.c 5.3
SCCS-vsn: usr.sbin/amd/rpcx/nfs_prot.h 5.3
SCCS-vsn: usr.sbin/amd/rpcx/nfs_prot_svc.c 5.3
SCCS-vsn: usr.sbin/amd/rpcx/nfs_prot_xdr.c 5.3
SCCS-vsn: usr.sbin/amd/rpcx/amq.x 5.3
SCCS-vsn: usr.sbin/amd/rpcx/amq.h 5.3
SCCS-vsn: usr.sbin/amd/config/newvers.sh 5.4
SCCS-vsn: usr.sbin/amd/config/Makefile.hpux 5.3
SCCS-vsn: usr.sbin/amd/config/Makefile.irix 5.3
SCCS-vsn: usr.sbin/amd/config/Makefile.config 5.3
SCCS-vsn: usr.sbin/amd/config/Makefile.bsd44 5.3
SCCS-vsn: usr.sbin/amd/config/os-type 5.4
SCCS-vsn: usr.sbin/amd/config/mount_aix.c 5.3
SCCS-vsn: usr.sbin/amd/config/mtab_bsd.c 5.3
SCCS-vsn: usr.sbin/amd/config/mtab_file.c 5.3
SCCS-vsn: usr.sbin/amd/config/mtab_ultrix.c 5.3
SCCS-vsn: usr.sbin/amd/config/os-acis43.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-aix3.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-aux.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-bsd44.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-concentrix.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-convex.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-defaults.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-dgux.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-hcx.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-hpux.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-irix.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-riscix.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-sos3.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-sos4.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-u2_2.h 5.3
SCCS-vsn: usr.sbin/amd/config/RELEASE 5.3
SCCS-vsn: usr.sbin/amd/config/os-u3_0.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-u4_0.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-utek.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-xinu43.h 5.3
SCCS-vsn: usr.sbin/amd/config/Configure 5.3
SCCS-vsn: usr.sbin/amd/config/arch 5.3
SCCS-vsn: usr.sbin/amd/config/Makefile.aix3 5.3
SCCS-vsn: usr.sbin/amd/config/misc-next.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-fpx4.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-hlh42.h 5.3
SCCS-vsn: usr.sbin/amd/config/os-next.h 5.4
SCCS-vsn: usr.sbin/amd/config/os-pyrOSx.h 5.3
SCCS-vsn: usr.sbin/amd/config/misc-irix.h 5.3
SCCS-vsn: usr.sbin/amd/config/misc-ultrix.h 5.3
SCCS-vsn: usr.sbin/amd/config/mount_irix.c 5.3
SCCS-vsn: usr.sbin/amd/config/os-umax43.h 5.3
SCCS-vsn: usr.sbin/amd/config/mtab_aix.c 5.3
SCCS-vsn: usr.sbin/amd/config/os-utx32.h 5.3
SCCS-vsn: usr.sbin/amd/config/misc-hpux.h 5.3
SCCS-vsn: usr.sbin/amd/config/misc-aix3.h 5.3
SCCS-vsn: usr.sbin/amd/amq/amq.8 5.5
SCCS-vsn: usr.sbin/amd/amq/amq.c 5.3
SCCS-vsn: usr.sbin/amd/amq/Makefile 5.4
SCCS-vsn: usr.sbin/amd/fsinfo/fsi_dict.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsi_gram.y 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsi_lex.l 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsi_data.h 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsi_util.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsinfo.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsinfo.h 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/wr_atab.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/wr_bparam.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/wr_dumpset.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/wr_exportfs.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/wr_fstab.c 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/Makefile 5.3
SCCS-vsn: usr.sbin/amd/fsinfo/fsi_analyze.c 5.3
SCCS-vsn: usr.sbin/amd/mk-amd-map/Makefile 5.4
SCCS-vsn: usr.sbin/amd/mk-amd-map/mk-amd-map.c 5.4
SCCS-vsn: usr.sbin/amd/text/amd.start.ex 5.3
SCCS-vsn: usr.sbin/amd/text/README 1.3
SCCS-vsn: usr.sbin/amd/text/COPYRIGHT 1.4
SCCS-vsn: usr.sbin/amd/text/INSTALL 1.3
SCCS-vsn: usr.sbin/amd/include/config.h 5.2
SCCS-vsn: usr.sbin/amd/include/uwait.h 5.3
SCCS-vsn: usr.sbin/amd/include/re.h 5.2
SCCS-vsn: usr.sbin/amd/include/remagic.h 5.2
SCCS-vsn: usr.sbin/amd/include/am.h 5.3
SCCS-vsn: usr.sbin/amd/include/fstype.h 5.3
SCCS-vsn: usr.sbin/amd/maps/a_net 5.3
SCCS-vsn: usr.sbin/amd/maps/a_master 5.3

+++ usr/src/usr.sbin/amd/amd/util.c
@@ -1,6 +1,4 @@
 /*
- * $Id: util.c,v 5.2.1.4 91/03/17 17:44:16 jsp Alpha $
- *
  * Copyright (c) 1990 Jan-Simon Pendry
  * Copyright (c) 1990 Imperial College of Science, Technology & Medicine
  * Copyright (c) 1990 The Regents of the University of California.
@@ -11,7 +9,10 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)util.c	5.2 (Berkeley) %G%
+ *	@(#)util.c	5.3 (Berkeley) %G%
+ *
+ * $Id: util.c,v 5.2.1.8 91/05/07 22:18:41 jsp Alpha $
+ *
  */
 
 /*
@@ -319,8 +320,9 @@ void mf_mounted(mf)
 mntfs *mf;
 {
 	int quoted;
+	int wasmounted = mf->mf_flags & MFF_MOUNTED;
 
-	if (!(mf->mf_flags & MFF_MOUNTED)) {
+	if (!wasmounted) {
 		/*
 		 * If this is a freshly mounted
 		 * filesystem then update the
@@ -342,10 +344,11 @@ mntfs *mf;
 	 * Log message
 	 */
 	quoted = strchr(mf->mf_info, ' ') != 0;
-	plog(XLOG_INFO, "%s%s%s mounted fstype %s on %s",
+	plog(XLOG_INFO, "%s%s%s %s fstype %s on %s",
 		quoted ? "\"" : "",
 		mf->mf_info,
 		quoted ? "\"" : "",
+		wasmounted ? "referenced" : "mounted",
 		mf->mf_ops->fs_type, mf->mf_mount);
 }
 
@@ -416,7 +419,8 @@ am_node *mp;
 	mf->mf_flags |= MFF_MOUNTING;
 	error = (*mf->mf_ops->mount_fs)(mp);
 	mf = mp->am_mnt;
-	mf->mf_flags &= ~MFF_MOUNTING;
+	if (error >= 0)
+		mf->mf_flags &= ~MFF_MOUNTING;
 	if (!error && !(mf->mf_ops->fs_flags & FS_MBACKGROUND)) {
 		/* ...but see ifs_mount */
 		am_mounted(mp);
@@ -475,7 +479,6 @@ am_node *mp;
  * TODO: Need a better strategy for handling errors
  */
 static int dofork(P_void);
-INLINE
 static int dofork()
 {
 	int pid;
@@ -502,13 +505,16 @@ int background()
 	if (pid == 0) {
 #ifdef DEBUG
 		dlog("backgrounded");
-#endif /* DEBUG */
+#endif
 		foreground = 0;
 	}
 
 	return pid;
 }
 
+/*
+ * Make all the directories in the path.
+ */
 int mkdirs P((char *path, int mode));
 int mkdirs(path, mode)
 char *path;
@@ -536,7 +542,7 @@ int mode;
 		} else {
 #ifdef DEBUG
 			dlog("mkdir(%s)", p2);
-#endif /* DEBUG */
+#endif
 		}
 		*sp = '/';
 	}
@@ -546,7 +552,7 @@ int mode;
 	} else {
 #ifdef DEBUG
 		dlog("mkdir(%s)", p2);
-#endif /* DEBUG */
+#endif
 	}
 
 #ifdef SUNOS4_WORKAROUND
@@ -572,6 +578,12 @@ int mode;
 		(stb.st_mode & S_IFMT) == S_IFDIR ? 0 : error_so_far;
 }
 
+/*
+ * Remove as many directories in the path as possible.
+ * Give up if the directory doesn't appear to have
+ * been created by Amd (not mode dr-x) or an rmdir
+ * fails for any reason.
+ */
 void rmdirs P((char *dir));
 void rmdirs(dir)
 char *dir;
@@ -588,13 +600,16 @@ char *dir;
 		 */
 		if (stat(xdp, &stb) == 0 && (stb.st_mode & 0200) == 0) {
 			if (rmdir(xdp) < 0) {
-				if (errno != ENOTEMPTY && errno != EBUSY && errno != EEXIST)
+				if (errno != ENOTEMPTY &&
+				    errno != EBUSY &&
+				    errno != EEXIST &&
+				    errno != EINVAL)
 					plog(XLOG_ERROR, "rmdir(%s): %m", xdp);
 				break;
 			} else {
 #ifdef DEBUG
 				dlog("rmdir(%s)", xdp);
-#endif /* DEBUG */
+#endif
 			}
 		} else {
 			break;
